feed_forward_vqgan_clip_/requirements.txt
torch==1.11.0
taming-transformers-rom1504==0.0.6
clip-anytorch==2.2.0
clize==4.2.0
einops==0.3.0
ftfy==6.0.3
fsspec==2021.8.1
imageio==2.9.0
kornia==0.5.10
torchvision==0.12.0
numpy==1.21.1
omegaconf==2.1.1
pandas==1.3.3
pytorch-lightning==1.4.7
requests==2.26.0
tensorboard==2.6.0
Pillow==9.0.1
torchmetrics==0.5.1
open_clip_torch==1.2.0
x-transformers==0.19.1